Moved UnlockWallet button up to page level
