+++ src/components/DelegateForm/DelegateForm.tsx
@@ -1,19 +1,15 @@
-import React, { Fragment, useState, useCallback } from 'react';
+import React, { useState, useCallback } from 'react';
 import { Card, CardTitle, CardContent, Button, Input, Spacer } from "react-neu";
 import Box from 'components/BoxWithDisplay';
 import Split from "components/Split";
-import UnlockWalletModal from "components/UnlockWalletModal";
 import { validateAddress } from 'utils';
-import { useWallet } from "use-wallet";
 
 import useGovernance from "hooks/useGovernance";
 import useFarming from "hooks/useFarming";
 
 export const DelegateForm: React.FC = () => {
   const [delegatee, setDelegatee] = useState('');
-  const { account } = useWallet();
 
-  const [unlockModalIsOpen, setUnlockModalIsOpen] = useState(false);
   const {
     isDelegated,
     onDelegateStaked,
@@ -35,53 +31,33 @@ export const DelegateForm: React.FC = () => {
 
   const onChange = (e: any): void => setDelegatee(e.target.value);
 
-  const handleDismissUnlockModal = useCallback(() => {
-    setUnlockModalIsOpen(false);
-  }, [setUnlockModalIsOpen]);
-
-  const handleUnlockWalletClick = useCallback(() => {
-    setUnlockModalIsOpen(true);
-  }, [setUnlockModalIsOpen])
-
-  if(account) {
-    return (
-      <Card>
-        <CardTitle text="Delegate Vote" />
-        <CardContent>
-          <Box display="grid" alignItems="center" paddingLeft={4} paddingRight={4} paddingBottom={1} row>
-            <Input onChange={onChange} placeholder="Delegate to..."></Input>
-            <Spacer />
-            <Split>
-              <Button
-                full
-                text="Delegate"
-                variant="secondary" 
-                onClick={isStaked ? handleOnDelegateStaked : handleOnDelegateUnstaked}
-                disabled={!validateAddress(delegatee)}
-                />
-              <Button
-                full
-                text="Remove Delegation"
-                variant="secondary"
-                onClick={onRemoveDelegation}
-                disabled={!isDelegated}/>
-            </Split>
-          </Box>
-        </CardContent>
-      </Card>
-    );
-  }
-  else {
-    return(
-      // TODO This should become a self-contained component shared across codebase
-      <Fragment>
-        <Box row justifyContent="center">
-          <Button onClick={handleUnlockWalletClick} text="Unlock wallet to display proposals" variant="secondary" />
+  return (
+    <Card>
+      <CardTitle text="Delegate Vote" />
+      <CardContent>
+        <Box display="grid" alignItems="center" paddingLeft={4} paddingRight={4} paddingBottom={1} row>
+          <Input onChange={onChange} placeholder="Delegate to..."></Input>
+          <Spacer />
+          <Split>
+            <Button
+              full
+              text="Delegate"
+              variant="secondary" 
+              onClick={isStaked ? handleOnDelegateStaked : handleOnDelegateUnstaked}
+              disabled={!validateAddress(delegatee)}
+              />
+            <Button
+              full
+              text="Remove Delegation"
+              variant="secondary"
+              onClick={onRemoveDelegation}
+              disabled={!isDelegated}/>
+          </Split>
         </Box>
-        <UnlockWalletModal isOpen={unlockModalIsOpen} onDismiss={handleDismissUnlockModal} />
-      </Fragment>
-    )
-  }
+      </CardContent>
+    </Card>
+  );
+  
 }
 
 export default DelegateForm;
