+++ src/views/Delegate/Delegate.tsx
@@ -1,16 +1,41 @@
-import React from 'react';
-import { Container, Spacer } from "react-neu";
+import React, {  useState, useCallback } from 'react';
+import { Container, Spacer, Button } from "react-neu";
 import Page from 'components/Page';
 import PageHeader from "components/PageHeader";
 import DelegateForm from "components/DelegateForm";
+import Box from 'components/BoxWithDisplay';
+import UnlockWalletModal from "components/UnlockWalletModal";
+import { useWallet } from "use-wallet";
+
 
 const Delegate = () => {
+  const { account } = useWallet();
+  const [unlockModalIsOpen, setUnlockModalIsOpen] = useState(false);
+
+  const handleDismissUnlockModal = useCallback(() => {
+    setUnlockModalIsOpen(false);
+  }, [setUnlockModalIsOpen]);
+
+  const handleUnlockWalletClick = useCallback(() => {
+    setUnlockModalIsOpen(true);
+  }, [setUnlockModalIsOpen])
+
   return (
     <Page>
       <Container>
         <PageHeader icon="📜" title="Delegate" />
         <Spacer size="md" />
-        <DelegateForm />
+        {account
+          ? <DelegateForm />
+          : (
+            <>
+              <Box row justifyContent="center">
+                <Button onClick={handleUnlockWalletClick} text="Unlock wallet to display proposals" variant="secondary" />
+              </Box>
+              <UnlockWalletModal isOpen={unlockModalIsOpen} onDismiss={handleDismissUnlockModal} />
+            </>
+          )
+        }
       </Container>
     </Page>
   );
